requirements.txt
Django==1.8.7
django-tastypie==0.12.2
djangorestframework==3.3.1
djangular==0.2.7
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.10.0
wheel==0.24.0
